Add files via upload
Binary search is one of the most common algorithms for finding items in sequential data types such as a list. Write a function in Python that takes a list of integers, and a specific integer and uses binary search to find if the desired integer is present in the list. If the integer is present, the function will return True, otherwise, it will return False. Print the result value and the location of result value as well.